in spel-katalog-batch, split up lua api across multiple rust modules
